requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.3.0
chardet==4.0.0
colormath==3.0.0
colour-science==0.3.16
discord.py==1.7.3
idna==3.3
imageio==2.13.5
joblib==1.1.0
multidict==5.2.0
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dotenv==0.19.2
PyWavelets==1.2.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
typing-extensions==4.0.1
yarl==1.7.2
